Updated Entities, changed Database in powershell script

+++ IkarusEntities/Address.cs
@@ -12,7 +12,7 @@ namespace IkarusEntities
             Customer = new HashSet<Customer>();
         }
 
-        public long AddressId { get; set; }
+        public int AddressId { get; set; }
         public string Address1 { get; set; }
         public string Address2 { get; set; }
         public string City { get; set; }
@@ -20,8 +20,8 @@ namespace IkarusEntities
         public DateTime DateCreated { get; set; }
         public DateTime? DateModified { get; set; }
         public bool? IsDeleted { get; set; }
-        public long AddressTypeId { get; set; }
-        public long CreatedByUserId { get; set; }
+        public int AddressTypeId { get; set; }
+        public int CreatedByUserId { get; set; }
 
         public AddressType AddressType { get; set; }
         public UserInfo CreatedByUser { get; set; }
