+++ IkarusEntities/CaseContact.cs
@@ -6,8 +6,8 @@ namespace IkarusEntities
     public partial class CaseContact
     {
         public long CaseContactId { get; set; }
-        public long CaseId { get; set; }
-        public long Contact { get; set; }
+        public int CaseId { get; set; }
+        public int Contact { get; set; }
 
         public CaseInfo Case { get; set; }
         public Contact ContactNavigation { get; set; }
