+++ IkarusEntities/ClientType.cs
@@ -10,12 +10,12 @@ namespace IkarusEntities
             Client = new HashSet<Client>();
         }
 
-        public long ClientTypeId { get; set; }
+        public int ClientTypeId { get; set; }
         public string ClientTypeName { get; set; }
         public DateTime DateCreated { get; set; }
         public DateTime? DateModified { get; set; }
         public bool? IsDeleted { get; set; }
-        public long? CustomerId { get; set; }
+        public int? CustomerId { get; set; }
 
         public Customer Customer { get; set; }
         public ICollection<Client> Client { get; set; }
