+++ IkarusEntities/Conversation.cs
@@ -11,9 +11,9 @@ namespace IkarusEntities
             Participant = new HashSet<Participant>();
         }
 
-        public long ConversationId { get; set; }
+        public int ConversationId { get; set; }
         public string ConversationName { get; set; }
-        public long? UserId { get; set; }
+        public int? UserId { get; set; }
 
         public UserInfo User { get; set; }
         public ICollection<Message> Message { get; set; }
