+++ IkarusEntities/Document.cs
@@ -5,13 +5,13 @@ namespace IkarusEntities
 {
     public partial class Document
     {
-        public long DocumentId { get; set; }
-        public long CaseId { get; set; }
-        public long DocumentCategoryId { get; set; }
-        public long FileTypeId { get; set; }
+        public int DocumentId { get; set; }
+        public int CaseId { get; set; }
+        public int DocumentCategoryId { get; set; }
+        public int FileTypeId { get; set; }
         public string DocumentPath { get; set; }
         public string DocumentContent { get; set; }
-        public long? CreatedByUserId { get; set; }
+        public int? CreatedByUserId { get; set; }
 
         public CaseInfo Case { get; set; }
         public UserInfo CreatedByUser { get; set; }
