+++ IkarusEntities/Message.cs
@@ -5,12 +5,12 @@ namespace IkarusEntities
 {
     public partial class Message
     {
-        public long MessageId { get; set; }
+        public int MessageId { get; set; }
         public string Message1 { get; set; }
         public DateTime DateCreated { get; set; }
         public DateTime? DateModified { get; set; }
-        public long? ConversationId { get; set; }
-        public long? CreatedByUserId { get; set; }
+        public int? ConversationId { get; set; }
+        public int? CreatedByUserId { get; set; }
 
         public Conversation Conversation { get; set; }
         public UserInfo CreatedByUser { get; set; }
