+++ IkarusEntities/Transaction.cs
@@ -5,12 +5,12 @@ namespace IkarusEntities
 {
     public partial class Transaction
     {
-        public long TransactionId { get; set; }
+        public int TransactionId { get; set; }
         public decimal Amount { get; set; }
         public DateTime DateCreated { get; set; }
-        public long PaymentGatewayId { get; set; }
-        public long PricingPackageId { get; set; }
-        public long CustomerId { get; set; }
+        public int PaymentGatewayId { get; set; }
+        public int PricingPackageId { get; set; }
+        public int CustomerId { get; set; }
 
         public Customer Customer { get; set; }
         public PaymentGateway PaymentGateway { get; set; }
