+++ IkarusEntities/UserCase.cs
@@ -5,10 +5,10 @@ namespace IkarusEntities
 {
     public partial class UserCase
     {
-        public long UserCaseId { get; set; }
+        public int UserCaseId { get; set; }
         public bool IsPrimary { get; set; }
-        public long CaseId { get; set; }
-        public long UserId { get; set; }
+        public int CaseId { get; set; }
+        public int UserId { get; set; }
 
         public CaseInfo Case { get; set; }
         public UserInfo User { get; set; }
