+++ IkarusEntities/UserMeeting.cs
@@ -5,9 +5,9 @@ namespace IkarusEntities
 {
     public partial class UserMeeting
     {
-        public long UserMeetingId { get; set; }
-        public long MeetingId { get; set; }
-        public long UserId { get; set; }
+        public int UserMeetingId { get; set; }
+        public int MeetingId { get; set; }
+        public int UserId { get; set; }
 
         public Meeting Meeting { get; set; }
         public UserInfo User { get; set; }
